requirements.txt
# Interface Inventory System - Python Dependencies
groq>=0.4.0
langgraph>=0.0.40
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
